+++ frontend/cstar_perf/frontend/static/js/schedule.js
@@ -8,10 +8,22 @@ var addRevisionDiv = function(animate){
     var revision_id = 'revision-'+schedule.n_revisions;
     var template = "<div id='{revision_id}' class='revision'><legend>Test Revisions<a id='remove-{revision_id}' class='pull-right remove-revision'><span class='glyphicon" +
         "                  glyphicon-remove'></span></a></legend>" +
+
+        "      <div class='form-group'>" +
+        "        <label class='col-md-4 control-label' for='{revision_id}-product'>Product</label>" +
+        "        <div class='col-md-8'>" +
+        "          <select id='{revision_id}-product' " +
+        "                  class='product-select form-control' required>" +
+        "            <option value='cassandra'>Cassandra</option>" +
+        "            <option value='dse'>DSE</option>" +
+        "          </select>" +
+        "        </div>" +
+        "      </div>" +
+
         "      <div class='form-group'>" +
         "        <label class='col-md-4 control-label' for='{revision_id}-refspec'>Revision</label>  " +
         "        <div class='col-md-8'>" +"" +
-        "          <input id='{revision_id}-refspec' type='text' placeholder='Git branch, tag, or commit id' class='refspec form-control input-md' required>" +
+        "          <input id='{revision_id}-refspec' type='text' placeholder='Git branch, tag, commit id or DSE version' class='refspec form-control input-md' required>" +
         "        </div>" +
         "      </div>" +
         "" +
@@ -64,7 +76,7 @@ var addRevisionDiv = function(animate){
         "      </div>" +
         "    </div>";
     var newDiv = $(template.format({revision:schedule.n_revisions, revision_id:revision_id}));
-    if (animate) 
+    if (animate)
         newDiv.hide();
     $("#schedule-revisions").append(newDiv);
     if (animate)
@@ -210,12 +222,13 @@ var createJob = function() {
         cluster: $("#cluster").val(),
         num_nodes: $("#numnodes").val(),
     }
-    
+
     //Revisions:
     job.revisions = [];
     $("#schedule-revisions div.revision").each(function(i, revision) {
         revision = $(revision);
         job.revisions[i] = {
+            product: revision.find(".product-select").val(),
             revision: revision.find(".refspec").val(),
             label: revision.find(".revision-label").val() ? revision.find(".revision-label").val() : null,
             yaml: revision.find(".yaml").val(),
@@ -256,7 +269,7 @@ var show_job_json = function() {
 var cloneExistingJob = function(job_id) {
     $.get("/api/tests/id/" + job_id, function(job) {
         test = job['test_definition'];
-        $("input#testname").val(test['title']); 
+        $("input#testname").val(test['title']);
         $("textarea#description").val(test['description']);
         $("select#cluster").val(test['cluster']);
         $("select#numnodes").val(test['num_nodes']);
@@ -265,6 +278,7 @@ var cloneExistingJob = function(job_id) {
             addRevisionDiv(false);
             var rev = i + 1;
             $("#revision-"+rev+"-refspec").val(revision['revision']);
+            $("#revision-"+rev+"-product").val(revision['product']);
             $("#revision-"+rev+"-label").val(revision['label']);
             $("#revision-"+rev+"-yaml").val(revision['yaml']);
             $("#revision-"+rev+"-env-vars").val(revision['env']);
@@ -317,7 +331,7 @@ var update_jvm_selections = function(callback) {
                 alert("Warning - cluster JVM selection changed")
             }
         });
-        if (callback != null) 
+        if (callback != null)
             callback();
     });
 }
@@ -356,7 +370,7 @@ $(document).ready(function() {
         addOperationDiv(false, 'stress', 'write n=19000000 -rate threads=50');
         addOperationDiv(false, 'stress', 'read n=19000000 -rate threads=50');
     }
-    
+
     //Validate form and submit:
     $("form#schedule-test").submit(function(e) {
         var job = createJob();
