+++ tool/cstar_perf/tool/benchmark.py
@@ -22,8 +22,9 @@ import yaml
 import sh
 import itertools
 
-# Import the default config first:
-import fab_cassandra as cstar
+# Import the default config first:r
+import fab_common as common
+import fab_dse as dse
 # Then import our cluster specific config:
 from cluster_config import config
 
@@ -37,11 +38,19 @@ sh.ErrorReturnCode.truncate_cap = 999999
 HOME = os.getenv('HOME')
 CASSANDRA_STRESS_PATH = os.path.expanduser("~/fab/stress/")
 CASSANDRA_STRESS_DEFAULT   = os.path.expanduser("~/fab/stress/default/tools/bin/cassandra-stress")
-CASSANDRA_NODETOOL = os.path.expanduser("~/fab/cassandra/bin/nodetool")
-CASSANDRA_CQLSH    = os.path.expanduser("~/fab/cassandra/bin/cqlsh")
 JAVA_HOME          = os.path.expanduser("~/fab/java")
 
 antcmd = sh.Command(os.path.join(HOME, 'fab/ant/bin/ant'))
+
+global nodetool_path, cqlsh_path
+
+def set_nodetool_path(path):
+    global nodetool_path
+    nodetool_path = path
+
+def set_cqlsh_path(path):
+    global cqlsh_path
+    cqlsh = path
 
 def bootstrap(cfg=None, destroy=False, leave_data=False, git_fetch=True):
     """Deploy and start cassandra on the cluster
@@ -54,7 +63,7 @@ def bootstrap(cfg=None, destroy=False, leave_data=False, git_fetch=True):
     Return the gid id of the branch checked out
     """
     if cfg is not None:
-        cstar.setup(cfg)
+        common.setup(cfg)
 
     # Parse yaml 
     if cfg.has_key('yaml'):
@@ -65,14 +74,14 @@ def bootstrap(cfg=None, destroy=False, leave_data=False, git_fetch=True):
             cass_yaml = {}
         if type(cass_yaml) is not dict:
             raise JobFailure('Invalid yaml, was expecting a dictionary: {cass_yaml}'.format(cass_yaml=cass_yaml))
-        cstar.config['yaml'] = cass_yaml
+        common.config['yaml'] = cass_yaml
     if cfg.has_key('options'):
         if cfg['options'] is not None:
-            cstar.config.update(cfg['options'])
-            del cstar.config['options']
+            common.config.update(cfg['options'])
+            del common.config['options']
 
     logger.info("### Config: ###")
-    pprint(cstar.config)
+    pprint(common.config)
 
     # leave_data settting can be set in the revision
     # configuration, or manually in the call to this function.
@@ -91,66 +100,80 @@ def bootstrap(cfg=None, destroy=False, leave_data=False, git_fetch=True):
 
     # Destroy cassandra deployment and data:
     if destroy:
-        execute(cstar.destroy, leave_data=leave_data)
-        execute(cstar.ensure_stopped)
+        execute(common.destroy, leave_data=leave_data)
+        execute(common.ensure_stopped)
     else:
         #Shutdown cleanly:
-        execute(cstar.stop)
-        execute(cstar.ensure_stopped)
+        execute(common.stop)
+        execute(common.ensure_stopped)
+
+    # dse setup and binaries download (local)
+    dse.setup(common.config)
+
+    product = dse if common.config['product'] == 'dse' else cstar
+    set_nodetool_path(os.path.join(product.get_bin_path(), 'nodetool'))
+    set_cqlsh_path(os.path.join(product.get_bin_path(), 'cqlsh'))
 
     # Bootstrap C* onto the cluster nodes, as well as the localhost,
     # so we have access to nodetool, stress etc - the local host will
     # not be added to the cluster unless it has a corresponding entry
     # in the cluster config:
-    hosts = list(cstar.fab.env['hosts'])
+    hosts = list(common.fab.env['hosts'])
     localhost = socket.gethostname().split(".")[0]
     if localhost not in [host.split(".")[0] for host in hosts]:
         # Use the local username for this host, as it may be different
         # than the cluster defined 'user' parameter:
         hosts += [getpass.getuser() + "@" + localhost]
     if not cfg.get('revision_override'):
-        with cstar.fab.settings(hosts=hosts):
-            git_ids = execute(cstar.bootstrap, git_fetch=git_fetch)
+        with common.fab.settings(hosts=hosts):
+            git_ids = execute(common.bootstrap, git_fetch=git_fetch)
     else:
+        # revision_override is only supported for the product cassandra
+        if product != 'cassandra':
+            raise ValueError("Cannot use revision_override for product: {}".format(
+                product.name))
         git_ids = {}
         default_hosts = set(hosts) - set(itertools.chain(*cfg['revision_override'].values()))
         print 'default version on {default_hosts}'.format(default_hosts=default_hosts)
-        with cstar.fab.settings(hosts=default_hosts):
-            git_ids.update(execute(cstar.bootstrap, git_fetch=git_fetch))
+        with common.fab.settings(hosts=default_hosts):
+            git_ids.update(execute(common.bootstrap, git_fetch=git_fetch))
         for override_revision, hosts_to_override in cfg['revision_override'].items():
 	    print '{revision} on {hosts_to_override}'.format(revision=override_revision, hosts_to_override=hosts_to_override)
-            with cstar.fab.settings(hosts=hosts_to_override):
-                git_ids.update(execute(cstar.bootstrap, git_fetch=git_fetch, revision_override=override_revision))
+            with common.fab.settings(hosts=hosts_to_override):
+                git_ids.update(execute(common.bootstrap, git_fetch=git_fetch, revision_override=override_revision))
 
-    overridden_host_versions = {}
-    for v, hs in cfg.get('revision_override', {}).items():
-        overridden_host_versions.update({h: v for h in hs})
-    expected_host_versions = dict({h: cfg['revision'] for h in hosts}, **overridden_host_versions)
-    expected_host_shas = {h: str(sh.git('--git-dir={home}/fab/cassandra.git'.format(home=HOME), 'rev-parse', v))
-                          for (h, v) in expected_host_versions.items()}
-    expected_host_shas = {h: v.strip() for (h, v) in expected_host_shas.items()}
+    if product.name == 'cassandra':
+        overridden_host_versions = {}
+        for v, hs in cfg.get('revision_override', {}).items():
+            overridden_host_versions.update({h: v for h in hs})
+        expected_host_versions = dict({h: cfg['revision'] for h in hosts}, **overridden_host_versions)
+        expected_host_shas = {h: str(sh.git('--git-dir={home}/fab/cassandra.git'.format(home=HOME), 'rev-parse', v))
+                              for (h, v) in expected_host_versions.items()}
+        expected_host_shas = {h: v.strip() for (h, v) in expected_host_shas.items()}
 
-    assert expected_host_shas == git_ids, 'expected: {}\ngot:{}'.format(expected_host_shas, git_ids)
+        assert expected_host_shas == git_ids, 'expected: {}\ngot:{}'.format(expected_host_shas, git_ids)
 
-    execute(cstar.start)
-    execute(cstar.ensure_running, hosts=[cstar.config['seeds'][0]])
+    execute(common.start)
+    execute(common.ensure_running, hosts=[common.config['seeds'][0]])
     time.sleep(30)
 
-    logger.info("Started cassandra on {n} nodes with git SHAs: {git_ids}".format(n=len(cstar.fab.env['hosts']), git_ids=git_ids))
+    logger.info("Started {product} on {n} nodes with git SHAs: {git_ids}".format(
+        product=product.name, n=len(common.fab.env['hosts']), git_ids=git_ids))
+    time.sleep(30)
     return git_ids
 
 def restart():
-    execute(cstar.stop)
-    execute(cstar.ensure_stopped)
-    execute(cstar.start)
-    execute(cstar.ensure_running)
+    execute(common.stop)
+    execute(common.ensure_stopped)
+    execute(common.start)
+    execute(common.ensure_running)
 
 def teardown(destroy=False, leave_data=False):
     if destroy:
-        execute(cstar.destroy, leave_data=leave_data)
+        execute(common.destroy, leave_data=leave_data)
     else:
-        execute(cstar.stop)
-        execute(cstar.ensure_stopped)
+        execute(common.stop)
+        execute(common.ensure_stopped)
 
 class NodetoolException(Exception):
     pass
@@ -160,8 +183,8 @@ def nodetool(cmd):
     
     Raises NodetoolException if we can't connect or another error occurs:
     """
-    cmd = "JAVA_HOME={JAVA_HOME} {CASSANDRA_NODETOOL} {cmd}".format(
-        JAVA_HOME=JAVA_HOME, CASSANDRA_NODETOOL=CASSANDRA_NODETOOL, cmd=cmd)
+    cmd = "JAVA_HOME={JAVA_HOME} {nodetool_path} {cmd}".format(
+        JAVA_HOME=JAVA_HOME, nodetool_path=nodetool_path, cmd=cmd)
     proc = subprocess.Popen(cmd, 
                             stdout=subprocess.PIPE,
                             stderr=subprocess.STDOUT,
@@ -183,15 +206,16 @@ def bash(script, nodes=None, user=None):
     if type(script) in (list, tuple):
         script = "\n".join(script)
     if nodes is None:
-        nodes = cstar.fab.env.hosts
+        nodes = common.fab.env.hosts
     if user is None:
-        user = cstar.fab.env.user
-    with cstar.fab.settings(user=user, hosts=nodes):
-        execute(cstar.bash, script)
+        user = common.fab.env.user
+    with common.fab.settings(user=user, hosts=nodes):
+        execute(common.bash, script)
 
 def cqlsh(script, node):
     """Run a cqlsh script on a node"""
-    proc = subprocess.Popen(shlex.split(CASSANDRA_CQLSH + " --no-color {host}".format(host=node)), 
+    cmd = "{cqlsh_path} --no-color {host}".format(cqlsh_path=cqlsh_path, host=node)
+    proc = subprocess.Popen(shlex.split(cmd),
                             stdout=subprocess.PIPE,
                             stdin=subprocess.PIPE,
                             stderr=subprocess.STDOUT)
@@ -295,7 +319,7 @@ def wait_for_compaction(nodes=None, check_interval=30, idle_confirmations=3,
                                 nodes=nodes, output=output))
 
     if nodes is None:
-        nodes = set(cstar.fab.env.hosts)
+        nodes = set(common.fab.env.hosts)
     else:
         nodes = set(nodes)
 
@@ -323,7 +347,7 @@ def set_device_read_ahead(read_ahead, devices=None):
     If devices argument is None, use the 'block_devices' setting from the cluster config."""
     if devices is None:
         devices = config['block_devices']
-    execute(cstar.set_device_read_ahead, read_ahead, devices)
+    execute(common.set_device_read_ahead, read_ahead, devices)
 
 def drop_page_cache():
     """Drop the page cache"""
@@ -432,19 +456,19 @@ def stress(cmd, revision_tag, stats=None, stress_revision=None):
 
 def retrieve_logs(local_directory):
     """Retrieve each node's logs to the given local directory."""
-    execute(cstar.copy_logs, local_directory=local_directory)
+    execute(common.copy_logs, local_directory=local_directory)
 
 def retrieve_fincore_logs(local_directory):
     """Retrieve each node's fincore logs to the given local directory."""
-    execute(cstar.copy_fincore_logs, local_directory=local_directory)
+    execute(common.copy_fincore_logs, local_directory=local_directory)
 
 def start_fincore_capture(interval=10):
     """Start linux-fincore monitoring of Cassandra data files on each node"""
-    execute(cstar.start_fincore_capture, interval=interval)
+    execute(common.start_fincore_capture, interval=interval)
 
 def stop_fincore_capture():
     """Stop linux-fincore monitoring"""
-    execute(cstar.stop_fincore_capture)
+    execute(common.stop_fincore_capture)
 
 def log_add_data(file, data):
     """Merge the dictionary data into the json log file root."""
